Remove CsvParserPlugin with LineDecoder, move classes to org.embulk.util.csv, and clean up dependencies
