+++ build.gradle
@@ -31,34 +31,9 @@ java {
 }
 
 dependencies {
-    compileOnly "org.embulk:embulk-api:0.10.32"
-    compileOnly "org.embulk:embulk-spi:0.10.32"
-
-    implementation "com.fasterxml.jackson.core:jackson-annotations:2.6.7"
-    implementation "com.fasterxml.jackson.core:jackson-core:2.6.7"
-    implementation "com.fasterxml.jackson.core:jackson-databind:2.6.7"
-    implementation "org.embulk:embulk-util-config:0.3.1"
-    implementation "org.embulk:embulk-util-file:0.1.3"
-    implementation "org.embulk:embulk-util-json:0.1.1"
-    implementation "org.embulk:embulk-util-text:0.1.1"
-    implementation "org.embulk:embulk-util-timestamp:0.2.1"
-
     testImplementation "junit:junit:4.12"
     testImplementation "org.embulk:embulk-api:0.10.32"
-    testImplementation "org.embulk:embulk-spi:0.10.32"
-    testImplementation "org.embulk:embulk-core:0.10.32"
-    testImplementation "org.embulk:embulk-core:0.10.32:tests"
-    testImplementation "org.embulk:embulk-deps:0.10.32"
-    testImplementation "org.embulk:embulk-junit4:0.10.32"
-
-    testImplementation "org.embulk:embulk-formatter-csv:0.10.32"
-    testImplementation "org.embulk:embulk-input-file:0.10.32"
-    testImplementation "org.embulk:embulk-output-file:0.10.32"
-    testImplementation "org.embulk:embulk-parser-csv:0.10.32"
-    testImplementation "org.embulk:embulk-guess-bzip2:0.10.32"
-    testImplementation "org.embulk:embulk-guess-csv:0.10.32"
-    testImplementation "org.embulk:embulk-guess-gzip:0.10.32"
-    testImplementation "org.embulk:embulk-guess-json:0.10.32"
+    testImplementation "org.embulk:embulk-util-text:0.1.1"
 }
 
 javadoc {
