+++ src/main/java/org/embulk/util/csv/RecordDoesNotHaveExpectedColumnException.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.embulk.parser.csv;
+package org.embulk.util.csv;
 
 public final class RecordDoesNotHaveExpectedColumnException extends Exception {
     public RecordDoesNotHaveExpectedColumnException() {
