+++ src/test/java/org/embulk/util/csv/TestWithRealFiles.java
@@ -28,12 +28,6 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import org.embulk.parser.csv.CsvTokenizer;
-import org.embulk.parser.csv.QuotedFieldLengthLimitExceededException;
-import org.embulk.parser.csv.RecordDoesNotHaveExpectedColumnException;
-import org.embulk.parser.csv.RecordHasUnexpectedRemainingColumnException;
-import org.embulk.parser.csv.UnexpectedCharacterAfterQuoteException;
-import org.embulk.parser.csv.UnexpectedEndOfLineInQuotedFieldException;
 import org.embulk.spi.Buffer;
 import org.embulk.spi.FileInput;
 import org.embulk.util.text.LineDecoder;
